pyproject.toml
[project]
name = "charcardview"
version = "0.1.0"
description = "Character card viewer for PNG files with embedded EXIF metadata"
requires-python = ">=3.12"
dependencies = [
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

